user-service/requirements.txt
flask
Flask_sqlalchemy
psycopg2-binary
